Move doorbell code to own file.
